requirements.txt
discord.py
requests
python-dotenv
matplotlib
plotly
kaleido
pandas